# contracts/api/openapi.yaml
openapi: 3.0.3
info:
  title: API consolidada (generado por scripts/generate-openapi.mjs)
  version: 1.0.0
  description: |
    Este archivo es generado automaticamente desde los api-contract.md
    de cada specs/<feature>/. NO editar a mano — perdera los cambios al
    proximo `npm run generate:openapi`.
    Features fuente: 001-catalogo-fuentes, 002-catalogo-readers, 003-diseno-y-ejecucion-procesos, 004-observabilidad-y-auditoria, 005-catalogo-conexiones, 006-programacion-procesos
servers:
  - url: /
paths:
...
  /api/reader-definitions:
    get:
      summary: Lista los readers configurados
      operationId: listReaderDefinitions
      responses:
        '200':
          description: OK
      x-source-feature: "002-catalogo-readers"
    post:
      summary: Crea un reader
      operationId: createReaderDefinition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReaderDefinition"
      responses:
        '200':
          description: OK
      x-source-feature: "002-catalogo-readers"
  /api/reader-definitions/{readerDefinitionId}:
    put:
      summary: Actualiza un reader
      operationId: updateReaderDefinition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReaderDefinition"
      responses:
        '200':
          description: OK
      x-source-feature: "002-catalogo-readers"
  /api/reader-definitions/{readerDefinitionId}/activation/{active}:
    post:
      summary: Activa o desactiva un reader
      operationId: setReaderDefinitionActive
      requestBody:
        required: false
        description: Sin cuerpo; el estado se indica en la ruta ({active}). Se referencia el recurso afectado por completitud del contrato.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReaderDefinition"
      responses:
        '200':
          description: OK
      x-source-feature: "002-catalogo-readers"
...
components:
  schemas:
...
    ReaderDefinition:
      type: object
      required:
        - id
        - name
        - readerType
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        readerType:
          type: string
        active:
          type: boolean
        configurationJson:
          type: string
...